GamesUPProject/gamesUP/src/test/java/com/gamesUP/gamesUP/Service/CategoryServiceTest.java
package com.gamesUP.gamesUP.Service;

import com.gamesUP.gamesUP.model.Category;
import com.gamesUP.gamesUP.Repository.CategoryRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Arrays;
import java.util.Optional;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.*;

class CategoryServiceTest {

    @Mock
    private CategoryRepository categoryRepository;

    @InjectMocks
    private CategoryService categoryService;

    private Category category;

    @BeforeEach
    void setUp() {
        // Initialise les mocks
        MockitoAnnotations.openMocks(this);
        category = new Category(1L, "Board Game");
    }

    @Test
    void testGetAllCategories() {
        when(categoryRepository.findAll()).thenReturn(Arrays.asList(category));

        var categories = categoryService.getAllCategories();

        assertEquals(1, categories.size());
        assertEquals("Board Game", categories.get(0).getType());
    }

    @Test
    void testGetCategoryById_Found() {
        when(categoryRepository.findById(1L)).thenReturn(Optional.of(category));

        Optional<Category> result = categoryService.getCategoryById(1L);

        assertTrue(result.isPresent());
        assertEquals("Board Game", result.get().getType());
    }

    @Test
    void testGetCategoryById_NotFound() {
        when(categoryRepository.findById(1L)).thenReturn(Optional.empty());

        Optional<Category> result = categoryService.getCategoryById(1L);

        assertFalse(result.isPresent());
    }

    @Test
    void testAddCategory() {
        when(categoryRepository.save(category)).thenReturn(category);

        Category result = categoryService.addCategory(category);

        assertEquals("Board Game", result.getType());
    }

    @Test
    void testUpdateCategory_Found() {
        when(categoryRepository.findById(1L)).thenReturn(Optional.of(category));
        when(categoryRepository.save(category)).thenReturn(category);

        Category updatedCategory = categoryService.updateCategory(1L, category);

        assertNotNull(updatedCategory);
        assertEquals("Board Game", updatedCategory.getType());
    }

    @Test
    void testUpdateCategory_NotFound() {
        when(categoryRepository.findById(1L)).thenReturn(Optional.empty());

        Category updatedCategory = categoryService.updateCategory(1L, category);

        assertNull(updatedCategory);
    }

    @Test
    void testDeleteCategory() {
        // Utilisation correcte de doNothing() pour les méthodes void
        doNothing().when(categoryRepository).deleteById(1L);

        categoryService.deleteCategory(1L);

        verify(categoryRepository, times(1)).deleteById(1L);
    }
}
